tools/datagnss/pyproject.toml
[project]
name = "datagnss_nano"
version = "0.0.1"
dependencies = ["pyserial"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
